% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgalaunay-package.R
\docType{package}
\name{cgalaunay-package}
\alias{cgalaunay}
\alias{cgalaunay-package}
\title{cgalaunay: Delaunay Triangulation of Planar Points with 'CGAL'}
\description{
Delaunay triangulation for points with the Computational Geometry Algorithms Library ('CGAL').
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
