% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delone-package.R
\docType{package}
\name{delone-package}
\alias{delone}
\alias{delone-package}
\title{delone: Delaunay Triangulation of Planar Points with 'CGAL'}
\description{
Delaunay triangulation for points with the Computational Geometry Algorithms Library ('CGAL').
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
